Fix host time sometimes skipping beats Closes #104
Turns out that the tick can be just at the end of the audio block
(ie, frame 511 of 512) and we will just barely miss it.
Because the tick is reset from host info on every audio block
(to always keep things in sync), we would miss the event because
of just 1 sample.
So give a little hand to make sure new beat is always triggered.

Signed-off-by: falkTX <falktx@falktx.com>