Sapphire v2.4.2
The following is a summary of changes between Sapphire versions 2.3.0 and 2.4.2.
See the Sapphire project change log for a more complete and detailed list:

https://github.com/cosinekitty/sapphire/blob/main/CHANGELOG.md

Added 5 new modules:
* Glee: a chaotic oscillator counterpart for Frolic
* Hiss: a 3D vector generator with unbiased directionality
* Nucleus: complex audio generator using physical simulation of electromagnetic particles
* Polynucleus: a version of Nucleus with vector-oriented polymorphic ports
* Tout (Tricorder OUTput): right-extender module for extracting a vector stream from Tricorder

Lots of improvements to Tricorder.
* Smoother animation
* Configurable axis rotation speed
* Other minor fixes and improvements

Bug fix: Elastika and TubeUnit now save and load Automatic Gain Control settings as part of the patch.

Added a context menu option in each attenuverter knob for a 10X low-sensitivity mode.
An orange dot appears on the knob to indicate low sensitivity.